Create scrapping thread class
Add new ScrapperThread class to scrapp websites (code from webscrapper) and throwing exceptions
Repair throwing exceptions in threads, now they work
Remove scrapping logic from WebScrapper.py
Chanage init a threads
Change logger error from testing "XDDDDD" to writing error message
